wip: adjust types
